requirements.txt
Django==3.2.15
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
drf-yasg==1.21.3
environs==9.5.0
psycopg2==2.9.3
gunicorn==20.1.0
pytest==7.0.1
pytest-django==4.5.2

